Add aliases and created-at fields to localUser
